refactor: request 타입별 메시지 처리를 CommandHandler 책임으로 위임
